Add files via upload
This bash script installs all dependencies, and then launches the nltk downloader to download the corpus.  You then need to click download on the window that pops up.

Should work on Ubuntu (python 3.5 should already be installed).  Can I have someone try this on Mac OS?  I want this script to work on any system.  

If you add more packages simply add to this script, and if you need to update packages just run this script and it will update as well.